Setup for testDouble-maybe? Quad launchfiles made not tested, python plot works
