+++ catkin_ws/bagfiles/time_size_dist_UAV0/UAV0_slash_dist_traversed.csv
@@ -1,0 +1,86 @@
+rosbagTimestamp,data
+1652967976909907154,0.0
+1652967977909790467,0.0
+1652967978909522537,2.83875501437e-07
+1652967979909474439,0.777553593328
+1652967980909712996,1.58892749459
+1652967981910276617,2.37862871583
+1652967982909671366,3.08519397811
+1652967983910096751,3.51105929915
+1652967984910083075,4.00977303491
+1652967985909790399,4.66226636824
+1652967986910343448,5.21617811036
+1652967987909574725,5.83763754304
+1652967988909730332,6.38878004355
+1652967989910477302,7.1103572024
+1652967990909522635,7.85541637964
+1652967991909463587,8.46832428049
+1652967992909551299,9.58135337837
+1652967993909513814,10.0378208246
+1652967994909822520,10.7155794767
+1652967995909494079,11.1464947245
+1652967996909544921,11.7440747126
+1652967997910239685,12.3402814004
+1652967998909502178,12.8072787651
+1652967999909491213,13.4703949885
+1652968000909756408,14.5308769686
+1652968001910517416,15.4599972995
+1652968002909627656,16.2553407341
+1652968003909507836,17.3043735422
+1652968004909679678,17.7572073582
+1652968005909570012,18.3277997084
+1652968006909546062,19.204901512
+1652968007909471283,19.5748260548
+1652968008912603945,20.1619189037
+1652968009910127668,20.4456055278
+1652968010909513813,21.1597559893
+1652968011909477202,21.4717981999
+1652968012909608944,22.4151343208
+1652968013909727624,22.8638638675
+1652968014910063517,23.2498021905
+1652968015909501104,23.8330838176
+1652968016909694012,24.4074234026
+1652968017910886847,25.1007905508
+1652968018909522075,26.1239748681
+1652968019909500943,26.657142832
+1652968020909640092,27.1243365757
+1652968021910548192,27.3617501629
+1652968022909751489,27.9201633553
+1652968023909497548,28.4264318964
+1652968024909566460,29.2020361864
+1652968025910096852,29.7106358077
+1652968026909527514,30.2287703642
+1652968027909945683,30.9162099317
+1652968028909544903,31.9178264033
+1652968029912062585,32.9163456049
+1652968030909971391,33.9286678399
+1652968031909526482,34.9664785352
+1652968032909601750,35.9570890806
+1652968033910272919,36.8522203036
+1652968034910058193,37.4492187356
+1652968035909563820,38.3094127432
+1652968036909566805,38.6608009847
+1652968037910180418,39.3547128968
+1652968038910523342,40.3530565032
+1652968039909840979,41.0989110561
+1652968040909698349,41.6638176902
+1652968041910282116,42.2884009292
+1652968042911000384,43.1568976074
+1652968043909615093,43.3129403778
+1652968044909547513,44.0139505714
+1652968045910085606,44.7849316198
+1652968046910076322,45.6987457148
+1652968047909572307,46.4913595459
+1652968048909564399,46.9498277684
+1652968049909614636,47.1503211397
+1652968050909575340,48.0323546013
+1652968051909482349,48.962532634
+1652968052909560524,49.9302351816
+1652968053909618537,50.6888646861
+1652968054909512102,51.1185332741
+1652968055910320579,51.4932558667
+1652968056909606050,51.5783266461
+1652968057910410516,51.5930706547
+1652968058910137229,51.595629984
+1652968059909588190,51.5957186502
+1652968060909547950,51.5957613618
