+++ catkin_ws/src/FUEL/fuel_planner/exploration_manager/launch/algorithm_MULTI.xml
@@ -14,6 +14,7 @@
   <arg name="sensor_pose_topic"/>
   <arg name="depth_topic"/>
   <arg name="cloud_topic"/>
+  <arg name="mapmerge_topic"/>
   
 
   <arg name="cx"/>
@@ -35,6 +36,7 @@
     <remap from ="/map_ros/pose"   to = "$(arg sensor_pose_topic)"/> 
     <remap from ="/map_ros/depth" to = "$(arg depth_topic)"/>
     <remap from ="/map_ros/cloud" to="$(arg cloud_topic)"/>
+    <remap from ="/MergedMap" to="$(arg mapmerge_topic)"/>
 
     <!-- for multi drone system -->
     <remap from ="/waypoint_generator/waypoints" to="/$(arg drone_index)/waypoint_generator/waypoints" />
