+++ catkin_ws/src/FUEL/fuel_planner/exploration_manager/launch/exploration_TRIPLE.launch
@@ -76,6 +76,8 @@
     <!-- subscribe ONLY TO ONE of the two topics -->
     <arg name="depth_topic" value="/$(arg ns0)/pcl_render_node/depth"/>
     <arg name="cloud_topic" value="/$(arg ns0)/pcl_render_node/cloud"/>
+
+    <arg name="mapmerge_topic" value="/$(arg ns0)/MergedMap"/>
     
 
     <!-- intrinsic params of the depth camera -->
@@ -152,6 +154,8 @@
     <!-- subscribe ONLY TO ONE of the two topics -->
     <arg name="depth_topic" value="/$(arg ns1)/pcl_render_node/depth"/>
     <arg name="cloud_topic" value="/$(arg ns1)/pcl_render_node/cloud"/>
+
+    <arg name="mapmerge_topic" value="/$(arg ns0)/MergedMap"/>
 
     <!-- intrinsic params of the depth camera -->
     <arg name="cx" value="321.04638671875"/>
@@ -230,6 +234,8 @@
     <!-- subscribe ONLY TO ONE of the two topics -->
     <arg name="depth_topic" value="/$(arg ns2)/pcl_render_node/depth"/>
     <arg name="cloud_topic" value="/$(arg ns2)/pcl_render_node/cloud"/>
+
+    <arg name="mapmerge_topic" value="/$(arg ns0)/MergedMap"/>
 
     <!-- intrinsic params of the depth camera -->
     <arg name="cx" value="321.04638671875"/>
