+++ catkin_ws/src/FUEL/fuel_planner/utils/lkh_tsp_solver/resource/UAV0/single.tsp
@@ -1,11 +1,10 @@
 NAME : single
 TYPE : ATSP
-DIMENSION : 4
+DIMENSION : 3
 EDGE_WEIGHT_TYPE : EXPLICIT
 EDGE_WEIGHT_FORMAT : FULL_MATRIX
 EDGE_WEIGHT_SECTION
-0 50338 139 127 
-0 0 396 426 
-0 396 0 57 
-0 426 57 0 
+0 214 208 
+0 0 131 
+0 131 0 
 EOF
