+++ catkin_ws/src/FUEL/fuel_planner/utils/lkh_tsp_solver/resource/UAV1/single.tsp
@@ -1,10 +1,18 @@
 NAME : single
 TYPE : ATSP
-DIMENSION : 3
+DIMENSION : 11
 EDGE_WEIGHT_TYPE : EXPLICIT
 EDGE_WEIGHT_FORMAT : FULL_MATRIX
 EDGE_WEIGHT_SECTION
-0 588 109 
-0 0 244 
-0 244 0 
+0 494 486 607 722 942 904 463 328 452 432 
+0 0 162 160 341 494 507 352 416 224 165 
+0 162 0 185 267 384 447 316 365 62 275 
+0 160 185 0 221 341 385 414 466 247 193 
+0 341 267 221 0 200 163 503 559 262 278 
+0 494 384 341 200 0 292 710 775 455 437 
+0 507 447 385 163 292 0 644 700 458 454 
+0 352 316 414 503 710 644 0 102 268 222 
+0 416 365 466 559 775 700 102 0 319 282 
+0 224 62 247 262 455 458 268 319 0 213 
+0 165 275 193 278 437 454 222 282 213 0 
 EOF
