+++ catkin_ws/src/FUEL/fuel_planner/utils/lkh_tsp_solver/resource/UAV1/single.txt
@@ -1,11 +1,19 @@
-NAME : single.353.tour
-COMMENT : Length = 353
-COMMENT : Found by LKH [Keld Helsgaun] Wed May 18 15:44:14 2022
+NAME : single.1925.tour
+COMMENT : Length = 1925
+COMMENT : Found by LKH [Keld Helsgaun] Thu May 19 11:52:12 2022
 TYPE : TOUR
-DIMENSION : 3
+DIMENSION : 11
 TOUR_SECTION
 1
+9
+8
+11
+10
 3
 2
+4
+5
+7
+6
 -1
 EOF
