+++ catkin_ws/src/autossar/coordination/launch/coordination.launch
@@ -37,8 +37,8 @@
 
 
 <!-- Start mapmerge utils -->
-    <!-- <node pkg="mapmerge" name="mapmerge" type="mapmerge" output="screen" ns="$(arg ns0)"> -->
-    <node pkg="mapmerge" name="mapmerge" type="mapmerge" ns="$(arg ns0)">
+    <!-- <node pkg="mapmerge" name="mapmerge" type="mapmerge" ns="$(arg ns0)"> -->
+    <node pkg="mapmerge" name="mapmerge" type="mapmerge" output="screen" ns="$(arg ns0)">
         <param name="otherUAV0" value="/$(arg ns1)"/>
         <param name="otherUAV1" value="/$(arg ns2)"/>
     </node>
@@ -54,17 +54,25 @@
     </node> -->
 
 <!-- Start coordination algorithm -->
-    <!-- <node pkg="coordination" name="coordination" type="coordination" output="screen" ns="$(arg ns0)">
+    <node pkg="coordination" name="coordination" type="coordination" output="screen" ns="$(arg ns0)">
         <param name="maxBattery" value="26"/>
     </node>
 
     <node pkg="coordination" name="coordination" type="coordination" output="screen" ns="$(arg ns1)">
         <param name="maxBattery" value="26"/>
-    </node> -->
+    </node>
 
     <!-- <node pkg="coordination" name="coordination" type="coordination" output="screen" ns="$(arg ns2)">
-        <param name="maxBattery" value="56"/> 
+        <param name="maxBattery" value="26"/> 
     </node> -->
 
+
+    <!-- <node pkg="rosbag" type="record" name="time_size_dist" args="record -O /UAV0/mapSize /UAV0/dist_traversed"/> -->
+    <node pkg="rosbag" type="record" name="time_size_dist_$(arg ns0)"
+    args="-O /home/malle/AutoSSAR/catkin_ws/bagfiles/time_size_dist_$(arg ns0) $(arg ns0)/mapSize $(arg ns0)/dist_traversed $(arg ns0)/within_range"/>
+    <node pkg="rosbag" type="record" name="time_size_dist_$(arg ns1)"
+    args="-O /home/malle/AutoSSAR/catkin_ws/bagfiles/time_size_dist_$(arg ns1) $(arg ns1)/mapSize $(arg ns1)/dist_traversed $(arg ns1)/within_range"/>
+    <!-- <node pkg="rosbag" type="record" name="time_size_dist_$(arg ns2)"
+    args="-O /home/malle/AutoSSAR/catkin_ws/bagfiles/time_size_dist_$(arg ns2) $(arg ns2)/mapSize $(arg ns2)/dist_traversed $(arg ns2)/within_range"/> -->
 
 </launch>
