+++ data/performanceDouble/performanceDouble.csv
@@ -1,0 +1,7 @@
+mapSize,dist,time,start,end,,mapSize,dist,time,start,end
+35012,19.6627522688,30.9039950370789,1652946129.22369,1652946160.12768,,35009,19.5048730122,32.2463893890381,1652946129.22369,1652946161.47008
+37365,25.7992703135,41.0247156620026,1652946627.97705,1652946669.00176,,34178,24.0027327324,39.1402797698975,1652946627.97705,1652946667.11733
+30710,20.1552428941,34.9651484489441,1652946756.90044,1652946791.86558,,28900,21.6430303902,35.8905665874481,1652946756.90044,1652946792.791
+36734,21.6975295761,32.8149580955505,1652948207.93568,1652948240.75064,,29663,19.1567208144,30.772123336792,1652948207.93568,1652948238.7078
+32925,20.4876032316,37.4624812602997,1652948402.49105,1652948439.95353,,36014,22.7538160034,37.5087115764618,1652948402.49105,1652948439.99976
+35102,19.869966193,33.8440096378326,1652953908.14702,1652953941.99103,,32222,19.1391489179,29.3666188716888,1652953908.14702,1652953937.51364
