+++ data/performanceDouble/performanceDoubleUAV0.csv
@@ -1,0 +1,7 @@
+mapSize,dist,time,start,end
+35012,19.6627522688,30.9039950370789,1652946129.22369,1652946160.12768
+37365,25.7992703135,41.0247156620026,1652946627.97705,1652946669.00176
+30710,20.1552428941,34.9651484489441,1652946756.90044,1652946791.86558
+36734,21.6975295761,32.8149580955505,1652948207.93568,1652948240.75064
+32925,20.4876032316,37.4624812602997,1652948402.49105,1652948439.95353
+35102,19.869966193,33.8440096378326,1652953908.14702,1652953941.99103
