+++ data/performanceDouble/performanceDoubleUAV1.csv
@@ -1,0 +1,7 @@
+mapSize,dist,time,start,end
+35009,19.5048730122,32.2463893890381,1652946129.22369,1652946161.47008
+34178,24.0027327324,39.1402797698975,1652946627.97705,1652946667.11733
+28900,21.6430303902,35.8905665874481,1652946756.90044,1652946792.791
+29663,19.1567208144,30.772123336792,1652948207.93568,1652948238.7078
+36014,22.7538160034,37.5087115764618,1652948402.49105,1652948439.99976
+32222,19.1391489179,29.3666188716888,1652953908.14702,1652953937.51364
